Refactor with feature-sliced structure (not optimal)
